Devex 864: finish splitting Case unit tests into separate files; add eslint rule
