+++ shared/src/business/entities/cases/Case.updateIrsPractitioner.test.js
@@ -3,8 +3,22 @@ const {
 } = require('../../test/createTestApplicationContext');
 const { Case } = require('./Case');
 const { IrsPractitioner } = require('../IrsPractitioner');
+const { MOCK_CASE } = require('../../../test/mockCase');
 
 describe('updateIrsPractitioner', () => {
+  let myCase;
+
+  beforeEach(() => {
+    myCase = new Case(
+      {
+        ...MOCK_CASE,
+        irsPractitioners: [{ name: 'Christopher Walken', userId: '123' }],
+        privatePractitioners: [{ name: 'Slim Shady', userId: '567' }],
+      },
+      { applicationContext },
+    );
+  });
+
   it('updates the given irsPractitioner on the case', () => {
     const caseToVerify = new Case(
       {
@@ -32,5 +46,30 @@ describe('updateIrsPractitioner', () => {
     expect(caseToVerify.irsPractitioners[0].email).toEqual(
       'irsPractitioner@example.com',
     );
+  });
+
+  it('updates a matching IRS practitioner found on the case', () => {
+    expect(myCase.irsPractitioners.length).toEqual(1);
+
+    myCase.updateIrsPractitioner({
+      name: 'Christopher Running',
+      userId: '123',
+    });
+
+    expect(myCase.irsPractitioners.length).toEqual(1);
+    expect(myCase.irsPractitioners[0]).toMatchObject({
+      name: 'Christopher Running',
+    });
+  });
+  it('updates nothing when provided object does not match', () => {
+    myCase.updateIrsPractitioner({
+      name: 'Slow Jog',
+      userId: '000-111-222',
+    });
+
+    expect(myCase.irsPractitioners.length).toEqual(1);
+    expect(myCase.irsPractitioners[0]).toMatchObject({
+      name: 'Christopher Walken',
+    });
   });
 });
